プログラムのmainへの移動
